Fix: wpcom-proxy-request was treating failed requests as successes
When the network connection is completely disconnected then
wpcom-proxy-request was returning a status code of zero, incorrectly
passing a test for success.

The referenced branch makes an explicit test for success instead of the
former looser test. This patch is here to run tests against the branch
and confirm the fix. Once we determine that all is well we can merge the
branch in the wpcom-proxy-request library, then update the version in
Calypso.