Fixed the "Redirect a site" form to remove the target url protocol and to have a prefix
So we have a problem that the site redirect allowed the user to add `http(s)://` in the target URL. The problem is mostly on the backend where the protocol is saved in one of the tables but not the other, so when the user tries to cancel the site redirect his site is left in a limbo (still having active redirect but the subscription is not active).

This PR fixes the "Redirect a site" page to include the protocol in front of the target URL and to automatically strip it from the field. It also updates the component to use classes and extend the React.Component.